added testcases

+++ AirlineProject/src/main/java/lti/project/backend/Pojos/FlightDetails.java
@@ -48,7 +48,7 @@ public class Flightdetails implements Serializable {
 	public Flightdetails() {
 	}
 
-	public long getFlightid() {
+	public int getFlightid() {
 		return this.flightid;
 	}
